requirements_openai.txt
streamlit==1.29.0
openai==0.28.0
python-dotenv==1.0.0
moviepy==1.0.3